Added Node and Variable highlight
