+++ addons/forge/editor/statescript/StatescriptGraphEditorDock.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using Gamesmiths.Forge.Godot.Resources.Statescript;
 using Godot;
 using GodotCollections = Godot.Collections;
@@ -400,6 +401,8 @@ public partial class StatescriptGraphEditorDock : EditorDock
 		};
 
 		_variablePanel.VariablesChanged += OnGraphVariablesChanged;
+		_variablePanel.VariableUndoRedoPerformed += OnVariableUndoRedoPerformed;
+		_variablePanel.VariableHighlightChanged += OnVariableHighlightChanged;
 		_splitContainer.AddChild(_variablePanel);
 
 		if (_undoRedo is not null)
@@ -504,176 +507,6 @@ public partial class StatescriptGraphEditorDock : EditorDock
 		AddChild(_addNodeDialog);
 
 		UpdateTheme();
-	}
-
-	private void OnFileMenuIdPressed(long id)
-	{
-		switch ((int)id)
-		{
-			case 0:
-				ShowNewStatescriptDialog();
-				break;
-
-			case 1:
-				ShowLoadStatescriptDialog();
-				break;
-
-			case 2:
-				OnSavePressed();
-				break;
-
-			case 3:
-				ShowSaveAsDialog();
-				break;
-
-			case 4:
-				CloseCurrentTab();
-				break;
-		}
-	}
-
-	private void ShowNewStatescriptDialog()
-	{
-		_newStatescriptDialog?.QueueFree();
-
-		_newStatescriptDialog = new AcceptDialog
-		{
-			Title = "Create Statescript",
-			Size = new Vector2I(400, 140),
-			Exclusive = true,
-		};
-
-		var vBox = new VBoxContainer { SizeFlagsHorizontal = SizeFlags.ExpandFill };
-		var pathRow = new HBoxContainer { SizeFlagsHorizontal = SizeFlags.ExpandFill };
-		vBox.AddChild(pathRow);
-
-		pathRow.AddChild(new Label { Text = "Path:", CustomMinimumSize = new Vector2(50, 0) });
-
-		_newStatescriptPathEdit = new LineEdit
-		{
-			Text = "res://new_statescript.tres",
-			SizeFlagsHorizontal = SizeFlags.ExpandFill,
-		};
-
-		pathRow.AddChild(_newStatescriptPathEdit);
-
-		_newStatescriptDialog.AddChild(vBox);
-		_newStatescriptDialog.Confirmed += OnNewStatescriptConfirmed;
-
-		AddChild(_newStatescriptDialog);
-		_newStatescriptDialog.PopupCentered();
-	}
-
-	private void OnNewStatescriptConfirmed()
-	{
-		if (_newStatescriptPathEdit is null)
-		{
-			return;
-		}
-
-		var path = _newStatescriptPathEdit.Text.Trim();
-		if (string.IsNullOrEmpty(path))
-		{
-			return;
-		}
-
-		if (!path.EndsWith(".tres", StringComparison.OrdinalIgnoreCase))
-		{
-			path += ".tres";
-		}
-
-		var graph = new StatescriptGraph();
-		graph.EnsureEntryNode();
-
-		graph.StatescriptName = path.GetFile().GetBaseName();
-
-		ResourceSaver.Save(graph, path);
-		EditorInterface.Singleton.GetResourceFilesystem().Scan();
-
-		graph = ResourceLoader.Load<StatescriptGraph>(path);
-		if (graph is not null)
-		{
-			OpenGraph(graph);
-		}
-
-		_newStatescriptDialog?.QueueFree();
-		_newStatescriptDialog = null;
-	}
-
-	private void ShowLoadStatescriptDialog()
-	{
-		var dialog = new EditorFileDialog
-		{
-			FileMode = FileDialog.FileModeEnum.OpenFile,
-			Title = "Load Statescript File",
-			Access = FileDialog.AccessEnum.Resources,
-		};
-
-		dialog.AddFilter("*.tres;StatescriptGraph");
-		dialog.FileSelected += path =>
-		{
-			Resource? graph = ResourceLoader.Load(path);
-			if (graph is StatescriptGraph statescriptGraph)
-			{
-				OpenGraph(statescriptGraph);
-			}
-			else
-			{
-				GD.PushWarning($"Failed to load StatescriptGraph from: {path}");
-			}
-
-			dialog.QueueFree();
-		};
-
-		dialog.Canceled += dialog.QueueFree;
-
-		AddChild(dialog);
-		dialog.PopupCentered(new Vector2I(700, 500));
-	}
-
-	private void ShowSaveAsDialog()
-	{
-		StatescriptGraph? graph = CurrentGraph;
-		if (graph is null)
-		{
-			return;
-		}
-
-		var dialog = new EditorFileDialog
-		{
-			FileMode = FileDialog.FileModeEnum.SaveFile,
-			Title = "Save Statescript As",
-			Access = FileDialog.AccessEnum.Resources,
-		};
-
-		dialog.AddFilter("*.tres", "Godot Resource");
-		dialog.FileSelected += path =>
-		{
-			if (_graphEdit is not null)
-			{
-				graph.ScrollOffset = _graphEdit.ScrollOffset;
-				graph.Zoom = _graphEdit.Zoom;
-				SyncVisualNodePositionsToGraph();
-				SyncConnectionsToCurrentGraph();
-			}
-
-			ResourceSaver.Save(graph, path);
-			EditorInterface.Singleton.GetResourceFilesystem().Scan();
-			GD.Print($"Statescript graph saved as: {path}");
-
-			StatescriptGraph? savedGraph = ResourceLoader.Load<StatescriptGraph>(path);
-			if (savedGraph is not null)
-			{
-				OpenGraph(savedGraph);
-			}
-
-			dialog.QueueFree();
-		};
-
-		dialog.Canceled += dialog.QueueFree;
-
-		AddChild(dialog);
-		dialog.PopupCentered(new Vector2I(700, 500));
 	}
 
 	private void UpdateTheme()
@@ -994,6 +827,55 @@ public partial class StatescriptGraphEditorDock : EditorDock
 		LoadGraphIntoEditor(graph);
 	}
 
+	private void OnVariableUndoRedoPerformed()
+	{
+		EnsureVariablesPanelVisible();
+	}
+
+	private void OnVariableHighlightChanged(string? variableName)
+	{
+		if (_graphEdit is null)
+		{
+			return;
+		}
+
+		foreach (Node child in _graphEdit.GetChildren())
+		{
+			if (child is StatescriptGraphNode graphNode)
+			{
+				graphNode.SetHighlightedVariable(variableName);
+			}
+		}
+	}
+
+	private void EnsureVariablesPanelVisible()
+	{
+		if (_variablePanel is null || _variablesToggleButton is null || _openTabs.Count == 0)
+		{
+			return;
+		}
+
+		if (_variablePanel.Visible)
+		{
+			return;
+		}
+
+		_variablePanel.Visible = true;
+		_variablesToggleButton.SetPressedNoSignal(true);
+
+		var current = _tabBar?.CurrentTab ?? -1;
+		if (current >= 0 && current < _openTabs.Count)
+		{
+			_openTabs[current].VariablesPanelOpen = true;
+		}
+
+		StatescriptGraph? graph = CurrentGraph;
+		if (graph is not null)
+		{
+			_variablePanel.SetGraph(graph);
+		}
+	}
+
 	private void OnFilesystemChanged()
 	{
 		for (var i = 0; i < _openTabs.Count; i++)
@@ -1037,194 +919,12 @@ public partial class StatescriptGraphEditorDock : EditorDock
 		}
 	}
 
-	private void OnBeginNodeMove()
+	private void OnGraphEditGuiInput(InputEvent @event)
 	{
-		if (_graphEdit is null)
+		if (@event is InputEventKey { Pressed: true, Keycode: Key.D, CtrlPressed: true })
 		{
-			return;
-		}
-
-		_preMovePositions.Clear();
-		foreach (Node child in _graphEdit.GetChildren())
-		{
-			if (child is StatescriptGraphNode { Selected: true } sgn)
-			{
-				_preMovePositions[sgn.Name] = sgn.PositionOffset;
-			}
-		}
-	}
-
-	private void OnEndNodeMove()
-	{
-		StatescriptGraph? graph = CurrentGraph;
-		if (graph is null || _graphEdit is null || _preMovePositions.Count == 0)
-		{
-			return;
-		}
-
-		var movedNodes = new GodotCollections.Dictionary<StringName, Vector2>();
-		var oldPositions = new GodotCollections.Dictionary<StringName, Vector2>();
-
-		foreach (Node child in _graphEdit.GetChildren())
-		{
-			if (child is not StatescriptGraphNode sgn || !_preMovePositions.TryGetValue(sgn.Name, out Vector2 oldPos))
-			{
-				continue;
-			}
-
-			Vector2 newPos = sgn.PositionOffset;
-			if (oldPos != newPos)
-			{
-				movedNodes[sgn.Name] = newPos;
-				oldPositions[sgn.Name] = oldPos;
-			}
-		}
-
-		_preMovePositions.Clear();
-
-		if (movedNodes.Count == 0)
-		{
-			return;
-		}
-
-		if (_undoRedo is not null)
-		{
-			_undoRedo.CreateAction("Move Statescript Node(s)", customContext: graph);
-			_undoRedo.AddDoMethod(this, MethodName.DoMoveNodes, graph, movedNodes);
-			_undoRedo.AddUndoMethod(this, MethodName.DoMoveNodes, graph, oldPositions);
-			_undoRedo.CommitAction(false);
-		}
-
-		SyncNodePositionsToResource(graph, movedNodes);
-	}
-
-	private void DoMoveNodes(
-		StatescriptGraph graph,
-		GodotCollections.Dictionary<StringName, Vector2> positions)
-	{
-		foreach (StatescriptNode node in graph.Nodes)
-		{
-			if (positions.TryGetValue(node.NodeId, out Vector2 pos))
-			{
-				node.PositionOffset = pos;
-			}
-		}
-
-		if (CurrentGraph == graph && _graphEdit is not null)
-		{
-			foreach (Node child in _graphEdit.GetChildren())
-			{
-				if (child is StatescriptGraphNode sgn && positions.TryGetValue(sgn.Name, out Vector2 pos))
-				{
-					sgn.PositionOffset = pos;
-				}
-			}
-		}
-	}
-
-	private void OnGraphEditPopupRequest(Vector2 atPosition)
-	{
-		if (CurrentGraph is null || _graphEdit is null || _addNodeDialog is null)
-		{
-			return;
-		}
-
-		ClearPendingConnection();
-
-		Vector2 graphPosition = (_graphEdit.ScrollOffset + atPosition) / _graphEdit.Zoom;
-		var screenPosition = (Vector2I)(_graphEdit.GetScreenPosition() + atPosition);
-
-		_addNodeDialog.ShowAtPosition(graphPosition, screenPosition);
-	}
-
-	private void OnConnectionToEmpty(StringName fromNode, long fromPort, Vector2 releasePosition)
-	{
-		if (CurrentGraph is null || _graphEdit is null || _addNodeDialog is null)
-		{
-			return;
-		}
-
-		_pendingConnectionNode = fromNode;
-		_pendingConnectionPort = (int)fromPort;
-		_pendingConnectionIsOutput = true;
-
-		Vector2 graphPosition = (_graphEdit.ScrollOffset + releasePosition) / _graphEdit.Zoom;
-		var screenPosition = (Vector2I)(_graphEdit.GetScreenPosition() + releasePosition);
-
-		_addNodeDialog.ShowAtPosition(graphPosition, screenPosition);
-	}
-
-	private void OnConnectionFromEmpty(StringName toNode, long toPort, Vector2 releasePosition)
-	{
-		if (CurrentGraph is null || _graphEdit is null || _addNodeDialog is null)
-		{
-			return;
-		}
-
-		_pendingConnectionNode = toNode;
-		_pendingConnectionPort = (int)toPort;
-		_pendingConnectionIsOutput = false;
-
-		Vector2 graphPosition = (_graphEdit.ScrollOffset + releasePosition) / _graphEdit.Zoom;
-		var screenPosition = (Vector2I)(_graphEdit.GetScreenPosition() + releasePosition);
-
-		_addNodeDialog.ShowAtPosition(graphPosition, screenPosition);
-	}
-
-	private void OnDialogNodeCreationRequested(
-		StatescriptNodeDiscovery.NodeTypeInfo? typeInfo,
-		StatescriptNodeType nodeType,
-		Vector2 position)
-	{
-		string newNodeId;
-
-		if (typeInfo is not null)
-		{
-			newNodeId = AddNodeAtPosition(nodeType, typeInfo.DisplayName, typeInfo.RuntimeTypeName, position);
-		}
-		else
-		{
-			newNodeId = AddNodeAtPosition(StatescriptNodeType.Exit, "Exit", string.Empty, position);
-		}
-
-		if (_pendingConnectionNode is not null && _graphEdit is not null)
-		{
-			if (_pendingConnectionIsOutput)
-			{
-				var inputPort = FindFirstEnabledInputPort(newNodeId);
-				if (inputPort >= 0)
-				{
-					OnConnectionRequest(
-						_pendingConnectionNode,
-						_pendingConnectionPort,
-						newNodeId,
-						inputPort);
-				}
-			}
-			else
-			{
-				var outputPort = FindFirstEnabledOutputPort(newNodeId);
-				if (outputPort >= 0)
-				{
-					OnConnectionRequest(
-						newNodeId,
-						outputPort,
-						_pendingConnectionNode,
-						_pendingConnectionPort);
-				}
-			}
-
-			ClearPendingConnection();
-		}
-	}
-
-	private void UndoAddNode(StatescriptGraph graph, StatescriptNode nodeResource)
-	{
-		graph.Nodes.Remove(nodeResource);
-
-		if (CurrentGraph == graph)
-		{
-			LoadGraphIntoEditor(graph);
+			DuplicateSelectedNodes();
+			GetViewport().SetInputAsHandled();
 		}
 	}
 
@@ -1276,398 +976,6 @@ public partial class StatescriptGraphEditorDock : EditorDock
 		return -1;
 	}
 
-	private void OnGraphEditGuiInput(InputEvent @event)
-	{
-		if (@event is InputEventKey { Pressed: true, Keycode: Key.D, CtrlPressed: true })
-		{
-			DuplicateSelectedNodes();
-			GetViewport().SetInputAsHandled();
-		}
-	}
-
-	private void DuplicateSelectedNodes()
-	{
-		StatescriptGraph? graph = CurrentGraph;
-		if (graph is null || _graphEdit is null)
-		{
-			return;
-		}
-
-		var selectedNodes = new List<StatescriptGraphNode>();
-
-		foreach (Node child in _graphEdit.GetChildren())
-		{
-			if (child is StatescriptGraphNode { Selected: true } statescriptNode
-				&& statescriptNode.NodeResource is not null
-				&& statescriptNode.NodeResource.NodeType != StatescriptNodeType.Entry)
-			{
-				selectedNodes.Add(statescriptNode);
-			}
-		}
-
-		if (selectedNodes.Count == 0)
-		{
-			return;
-		}
-
-		foreach (StatescriptGraphNode sgn in selectedNodes)
-		{
-			sgn.Selected = false;
-		}
-
-		var duplicatedIds = new Dictionary<string, string>();
-		const float offset = 40f;
-
-		foreach (StatescriptGraphNode sgn in selectedNodes)
-		{
-			StatescriptNode original = sgn.NodeResource!;
-			var newNodeId = $"node_{_nextNodeId++}";
-			duplicatedIds[original.NodeId] = newNodeId;
-
-			var duplicated = new StatescriptNode
-			{
-				NodeId = newNodeId,
-				Title = original.Title,
-				NodeType = original.NodeType,
-				RuntimeTypeName = original.RuntimeTypeName,
-				PositionOffset = original.PositionOffset + new Vector2(offset, offset),
-			};
-
-			foreach (KeyValuePair<string, Variant> kvp in original.CustomData)
-			{
-				duplicated.CustomData[kvp.Key] = kvp.Value;
-			}
-
-			foreach (StatescriptNodeProperty binding in original.PropertyBindings)
-			{
-				var newBinding = new StatescriptNodeProperty
-				{
-					Direction = binding.Direction,
-					PropertyIndex = binding.PropertyIndex,
-					Resolver = binding.Resolver is not null
-						? (StatescriptResolverResource)binding.Resolver.Duplicate(true)
-						: null,
-				};
-
-				duplicated.PropertyBindings.Add(newBinding);
-			}
-
-			graph.Nodes.Add(duplicated);
-
-			var graphNode = new StatescriptGraphNode();
-			_graphEdit.AddChild(graphNode);
-			graphNode.Initialize(duplicated, graph);
-			graphNode.Selected = true;
-		}
-
-		foreach (StatescriptConnection connection in graph.Connections)
-		{
-			if (duplicatedIds.TryGetValue(connection.FromNode, out var newFrom)
-				&& duplicatedIds.TryGetValue(connection.ToNode, out var newTo))
-			{
-				_graphEdit.ConnectNode(newFrom, connection.OutputPort, newTo, connection.InputPort);
-			}
-		}
-
-		SyncConnectionsToCurrentGraph();
-	}
-
-	private void OnConnectionRequest(StringName fromNode, long fromPort, StringName toNode, long toPort)
-	{
-		StatescriptGraph? graph = CurrentGraph;
-		if (graph is null || _graphEdit is null)
-		{
-			return;
-		}
-
-		if (_undoRedo is not null)
-		{
-			_undoRedo.CreateAction("Connect Statescript Nodes", customContext: graph);
-			_undoRedo.AddDoMethod(
-				this,
-				MethodName.DoConnect,
-				fromNode.ToString(),
-				(int)fromPort,
-				toNode.ToString(),
-				(int)toPort);
-			_undoRedo.AddUndoMethod(
-				this,
-				MethodName.UndoConnect,
-				fromNode.ToString(),
-				(int)fromPort,
-				toNode.ToString(),
-				(int)toPort);
-			_undoRedo.CommitAction();
-		}
-		else
-		{
-			DoConnect(fromNode.ToString(), (int)fromPort, toNode.ToString(), (int)toPort);
-		}
-	}
-
-	private void DoConnect(string fromNode, int fromPort, string toNode, int toPort)
-	{
-		_graphEdit?.ConnectNode(fromNode, fromPort, toNode, toPort);
-		SyncConnectionsToCurrentGraph();
-	}
-
-	private void UndoConnect(string fromNode, int fromPort, string toNode, int toPort)
-	{
-		_graphEdit?.DisconnectNode(fromNode, fromPort, toNode, toPort);
-		SyncConnectionsToCurrentGraph();
-	}
-
-	private void OnDisconnectionRequest(StringName fromNode, long fromPort, StringName toNode, long toPort)
-	{
-		StatescriptGraph? graph = CurrentGraph;
-		if (graph is null || _graphEdit is null)
-		{
-			return;
-		}
-
-		if (_undoRedo is not null)
-		{
-			_undoRedo.CreateAction("Disconnect Statescript Nodes", customContext: graph);
-			_undoRedo.AddDoMethod(
-				this,
-				MethodName.UndoConnect,
-				fromNode.ToString(),
-				(int)fromPort,
-				toNode.ToString(),
-				(int)toPort);
-			_undoRedo.AddUndoMethod(
-				this,
-				MethodName.DoConnect,
-				fromNode.ToString(),
-				(int)fromPort,
-				toNode.ToString(),
-				(int)toPort);
-			_undoRedo.CommitAction();
-		}
-		else
-		{
-			_graphEdit.DisconnectNode(fromNode, (int)fromPort, toNode, (int)toPort);
-			SyncConnectionsToCurrentGraph();
-		}
-	}
-
-	private void OnDeleteNodesRequest(GodotCollections.Array<StringName> deletedNodes)
-	{
-		StatescriptGraph? graph = CurrentGraph;
-		if (graph is null || _graphEdit is null)
-		{
-			return;
-		}
-
-		foreach (StringName nodeName in deletedNodes)
-		{
-			Node? child = _graphEdit.GetNodeOrNull(nodeName.ToString());
-
-			if (child is not StatescriptGraphNode graphNode)
-			{
-				continue;
-			}
-
-			if (graphNode.NodeResource?.NodeType == StatescriptNodeType.Entry)
-			{
-				GD.PushWarning("Cannot delete the Entry statescriptNode.");
-				continue;
-			}
-
-			if (graphNode.NodeResource is null)
-			{
-				continue;
-			}
-
-			var affectedConnections = new List<StatescriptConnection>();
-			foreach (GodotCollections.Dictionary connection in _graphEdit.GetConnectionList())
-			{
-				StringName from = connection["from_node"].AsStringName();
-				StringName to = connection["to_node"].AsStringName();
-
-				if (from == nodeName || to == nodeName)
-				{
-					affectedConnections.Add(new StatescriptConnection
-					{
-						FromNode = connection["from_node"].AsString(),
-						OutputPort = connection["from_port"].AsInt32(),
-						ToNode = connection["to_node"].AsString(),
-						InputPort = connection["to_port"].AsInt32(),
-					});
-				}
-			}
-
-			if (_undoRedo is not null)
-			{
-				_undoRedo.CreateAction("Delete Statescript Node", customContext: graph);
-				_undoRedo.AddDoMethod(
-					this,
-					MethodName.DoDeleteNode,
-					graph,
-					graphNode.NodeResource,
-					new GodotCollections.Array<StatescriptConnection>(affectedConnections));
-				_undoRedo.AddUndoMethod(
-					this,
-					MethodName.UndoDeleteNode,
-					graph,
-					graphNode.NodeResource,
-					new GodotCollections.Array<StatescriptConnection>(affectedConnections));
-				_undoRedo.CommitAction();
-			}
-			else
-			{
-				DoDeleteNode(
-					graph,
-					graphNode.NodeResource,
-					[.. affectedConnections]);
-			}
-		}
-	}
-
-	private void DoDeleteNode(
-		StatescriptGraph graph,
-		StatescriptNode nodeResource,
-		GodotCollections.Array<StatescriptConnection> affectedConnections)
-	{
-		if (_graphEdit is not null && CurrentGraph == graph)
-		{
-			foreach (StatescriptConnection connection in affectedConnections)
-			{
-				_graphEdit.DisconnectNode(
-					connection.FromNode,
-					connection.OutputPort,
-					connection.ToNode,
-					connection.InputPort);
-			}
-
-			Node? child = _graphEdit.GetNodeOrNull(nodeResource.NodeId);
-			child?.QueueFree();
-		}
-
-		graph.Nodes.Remove(nodeResource);
-		SyncConnectionsToCurrentGraph();
-	}
-
-	private void UndoDeleteNode(
-		StatescriptGraph graph,
-		StatescriptNode nodeResource,
-		GodotCollections.Array<StatescriptConnection> affectedConnections)
-	{
-		graph.Nodes.Add(nodeResource);
-
-		graph.Connections.AddRange(affectedConnections);
-
-		if (CurrentGraph == graph)
-		{
-			LoadGraphIntoEditor(graph);
-		}
-	}
-
-	private void OnDialogCanceled()
-	{
-		ClearPendingConnection();
-	}
-
-	private void ClearPendingConnection()
-	{
-		_pendingConnectionNode = null;
-		_pendingConnectionPort = 0;
-		_pendingConnectionIsOutput = false;
-	}
-
-	private void OnAddNodeButtonPressed()
-	{
-		if (CurrentGraph is null || _graphEdit is null || _addNodeDialog is null || _addNodeButton is null)
-		{
-			return;
-		}
-
-		ClearPendingConnection();
-
-		var screenPosition = (Vector2I)(_addNodeButton.GetScreenPosition() + new Vector2(0, _addNodeButton.Size.Y));
-
-		Vector2 centerPosition = (_graphEdit.ScrollOffset + (_graphEdit.Size / 2)) / _graphEdit.Zoom;
-
-		_addNodeDialog.ShowAtPosition(centerPosition, screenPosition);
-	}
-
-	private string AddNodeAtPosition(
-		StatescriptNodeType nodeType,
-		string title,
-		string runtimeTypeName,
-		Vector2 position)
-	{
-		StatescriptGraph? graph = CurrentGraph;
-		if (graph is null || _graphEdit is null)
-		{
-			return string.Empty;
-		}
-
-		var nodeId = $"node_{_nextNodeId++}";
-
-		var nodeResource = new StatescriptNode
-		{
-			NodeId = nodeId,
-			Title = title,
-			NodeType = nodeType,
-			RuntimeTypeName = runtimeTypeName,
-			PositionOffset = position,
-		};
-
-		if (_undoRedo is not null)
-		{
-			_undoRedo.CreateAction("Add Statescript Node", customContext: graph);
-			_undoRedo.AddDoMethod(this, MethodName.DoAddNode, graph, nodeResource);
-			_undoRedo.AddUndoMethod(this, MethodName.UndoAddNode, graph, nodeResource);
-			_undoRedo.CommitAction();
-		}
-		else
-		{
-			DoAddNode(graph, nodeResource);
-		}
-
-		return nodeId;
-	}
-
-	private void DoAddNode(StatescriptGraph graph, StatescriptNode nodeResource)
-	{
-		graph.Nodes.Add(nodeResource);
-
-		if (CurrentGraph == graph && _graphEdit is not null)
-		{
-			var graphNode = new StatescriptGraphNode();
-			_graphEdit.AddChild(graphNode);
-			graphNode.Initialize(nodeResource, graph);
-			graphNode.SetUndoRedo(_undoRedo);
-		}
-	}
-
-	private void OnSavePressed()
-	{
-		StatescriptGraph? graph = CurrentGraph;
-		if (graph is null || _graphEdit is null)
-		{
-			return;
-		}
-
-		graph.ScrollOffset = _graphEdit.ScrollOffset;
-		graph.Zoom = _graphEdit.Zoom;
-		SyncVisualNodePositionsToGraph();
-		SyncConnectionsToCurrentGraph();
-
-		if (string.IsNullOrEmpty(graph.ResourcePath))
-		{
-			ShowSaveAsDialog();
-			return;
-		}
-
-		ResourceSaver.Save(graph);
-		GD.Print($"Statescript graph saved: {graph.ResourcePath}");
-	}
-
-	/// <summary>
-	/// Internal record to track open tab state.
-	/// </summary>
 	private sealed class GraphTab
 	{
 		private string _cachedPath;
