+++ addons/forge/editor/statescript/StatescriptGraphNode.cs
@@ -30,6 +30,7 @@ public partial class StatescriptGraphNode : GraphNode
 	private static readonly Color _subgraphColor = new(0xc678ddff);
 	private static readonly Color _inputPropertyColor = new(0x61afefff);
 	private static readonly Color _outputVariableColor = new(0xe5c07bff);
+	private static readonly Color _highlightColor = new(0x56b6c2ff);
 
 	private readonly Dictionary<PropertySlotKey, NodeEditorProperty> _activeResolverEditors = [];
 
@@ -38,6 +39,8 @@ public partial class StatescriptGraphNode : GraphNode
 	private EditorUndoRedoManager? _undoRedo;
 	private bool _resizeConnected;
 	private float _widthBeforeResize;
+	private string? _highlightedVariableName;
+	private bool _isHighlighted;
 
 	/// <summary>
 	/// Raised when a property binding has been modified in the UI.
@@ -65,6 +68,18 @@ public partial class StatescriptGraphNode : GraphNode
 	public EditorUndoRedoManager? GetUndoRedo()
 	{
 		return _undoRedo;
+	}
+
+	/// <summary>
+	/// Updates the highlight state based on the given variable name.
+	/// </summary>
+	/// <param name="variableName">The variable name to highlight, or null to clear.</param>
+	public void SetHighlightedVariable(string? variableName)
+	{
+		_highlightedVariableName = variableName;
+		_isHighlighted = !string.IsNullOrEmpty(variableName) && ReferencesVariable(variableName!);
+		ApplyHighlightBorder();
+		UpdateChildHighlights();
 	}
 
 	/// <summary>
@@ -353,6 +368,7 @@ public partial class StatescriptGraphNode : GraphNode
 		};
 
 		sectionContainer.AddThemeColorOverride("font_color", color);
+
 		sectionContainer.FoldingChanged += isFolded =>
 		{
 			SetFoldStateWithUndo(foldKey, isFolded);
@@ -485,344 +501,6 @@ public partial class StatescriptGraphNode : GraphNode
 		NodeResource.CustomData[CustomWidthKey] = Variant.From(width);
 	}
 
-	private void AddInputPropertyRow(
-		StatescriptNodeDiscovery.InputPropertyInfo propInfo,
-		int index,
-		Control sectionContainer)
-	{
-		if (NodeResource is null)
-		{
-			return;
-		}
-
-		var container = new VBoxContainer { SizeFlagsHorizontal = SizeFlags.ExpandFill };
-		sectionContainer.AddChild(container);
-
-		var headerRow = new HBoxContainer { SizeFlagsHorizontal = SizeFlags.ExpandFill };
-		container.AddChild(headerRow);
-
-		var nameLabel = new Label
-		{
-			Text = propInfo.Label,
-			CustomMinimumSize = new Vector2(60, 0),
-		};
-
-		nameLabel.AddThemeColorOverride("font_color", _inputPropertyColor);
-		headerRow.AddChild(nameLabel);
-
-		List<Func<NodeEditorProperty>> resolverFactories =
-			StatescriptResolverRegistry.GetCompatibleFactories(propInfo.ExpectedType);
-
-		if (resolverFactories.Count == 0)
-		{
-			var errorLabel = new Label
-			{
-				Text = "No compatible resolvers.",
-			};
-
-			errorLabel.AddThemeColorOverride("font_color", Colors.Red);
-			headerRow.AddChild(errorLabel);
-			return;
-		}
-
-		var resolverDropdown = new OptionButton
-		{
-			SizeFlagsHorizontal = SizeFlags.ExpandFill,
-			CustomMinimumSize = new Vector2(80, 0),
-		};
-
-		foreach (Func<NodeEditorProperty> factory in resolverFactories)
-		{
-			using NodeEditorProperty temp = factory();
-			resolverDropdown.AddItem(temp.DisplayName);
-		}
-
-		StatescriptNodeProperty? binding = FindBinding(StatescriptPropertyDirection.Input, index);
-		var selectedIndex = 0;
-
-		if (binding?.Resolver is not null)
-		{
-			for (var i = 0; i < resolverFactories.Count; i++)
-			{
-				using NodeEditorProperty temp = resolverFactories[i]();
-
-				if (temp.ResolverTypeId == GetResolverTypeId(binding.Resolver))
-				{
-					selectedIndex = i;
-					break;
-				}
-			}
-		}
-		else
-		{
-			for (var i = 0; i < resolverFactories.Count; i++)
-			{
-				using NodeEditorProperty temp = resolverFactories[i]();
-
-				if (temp.ResolverTypeId == "Variant")
-				{
-					selectedIndex = i;
-					break;
-				}
-			}
-		}
-
-		resolverDropdown.Selected = selectedIndex;
-		headerRow.AddChild(resolverDropdown);
-
-		var editorContainer = new VBoxContainer { SizeFlagsHorizontal = SizeFlags.ExpandFill };
-		container.AddChild(editorContainer);
-
-		ShowResolverEditorUI(
-			resolverFactories[selectedIndex],
-			binding,
-			propInfo.ExpectedType,
-			editorContainer,
-			StatescriptPropertyDirection.Input,
-			index,
-			propInfo.IsArray);
-
-		resolverDropdown.ItemSelected += x =>
-		{
-			var key = new PropertySlotKey(StatescriptPropertyDirection.Input, index);
-
-			var oldResolver = FindBinding(StatescriptPropertyDirection.Input, index)?.Resolver?.Duplicate()
-				as StatescriptResolverResource;
-
-			if (_activeResolverEditors.TryGetValue(key, out NodeEditorProperty? old))
-			{
-				_activeResolverEditors.Remove(key);
-			}
-
-			ClearContainer(editorContainer);
-
-			if (NodeResource is null)
-			{
-				return;
-			}
-
-			ShowResolverEditorUI(
-				resolverFactories[(int)x],
-				null,
-				propInfo.ExpectedType,
-				editorContainer,
-				StatescriptPropertyDirection.Input,
-				index,
-				propInfo.IsArray);
-
-			if (_activeResolverEditors.TryGetValue(key, out NodeEditorProperty? editor))
-			{
-				SaveResolverEditor(editor, StatescriptPropertyDirection.Input, index);
-			}
-
-			StatescriptNodeProperty? updated = FindBinding(StatescriptPropertyDirection.Input, index);
-			var newResolver = updated?.Resolver?.Duplicate() as StatescriptResolverResource;
-
-			if (_undoRedo is not null)
-			{
-				_undoRedo.CreateAction("Change Resolver Type", customContext: _graph);
-				_undoRedo.AddDoMethod(
-					this,
-					MethodName.ApplyResolverBinding,
-					(int)StatescriptPropertyDirection.Input,
-					index,
-					newResolver ?? new StatescriptResolverResource());
-				_undoRedo.AddUndoMethod(
-					this,
-					MethodName.ApplyResolverBinding,
-					(int)StatescriptPropertyDirection.Input,
-					index,
-					oldResolver ?? new StatescriptResolverResource());
-				_undoRedo.CommitAction(false);
-			}
-
-			PropertyBindingChanged?.Invoke();
-			ResetSize();
-		};
-	}
-
-	private void AddOutputVariableRow(
-		StatescriptNodeDiscovery.OutputVariableInfo varInfo,
-		int index,
-		FoldableContainer sectionContainer)
-	{
-		if (NodeResource is null || _graph is null)
-		{
-			return;
-		}
-
-		var hBox = new HBoxContainer { SizeFlagsHorizontal = SizeFlags.ExpandFill };
-		sectionContainer.AddChild(hBox);
-
-		var nameLabel = new Label
-		{
-			Text = varInfo.Label,
-			CustomMinimumSize = new Vector2(60, 0),
-		};
-
-		nameLabel.AddThemeColorOverride("font_color", _outputVariableColor);
-		hBox.AddChild(nameLabel);
-
-		var variableDropdown = new OptionButton { SizeFlagsHorizontal = SizeFlags.ExpandFill };
-
-		foreach (StatescriptGraphVariable v in _graph.Variables)
-		{
-			variableDropdown.AddItem(v.VariableName);
-		}
-
-		StatescriptNodeProperty? binding = FindBinding(StatescriptPropertyDirection.Output, index);
-		var selectedIndex = 0;
-
-		if (binding?.Resolver is VariableResolverResource varRes
-			&& !string.IsNullOrEmpty(varRes.VariableName))
-		{
-			for (var i = 0; i < _graph.Variables.Count; i++)
-			{
-				if (_graph.Variables[i].VariableName == varRes.VariableName)
-				{
-					selectedIndex = i;
-					break;
-				}
-			}
-		}
-
-		if (_graph.Variables.Count > 0)
-		{
-			variableDropdown.Selected = selectedIndex;
-
-			if (binding is null)
-			{
-				var variableName = _graph.Variables[selectedIndex].VariableName;
-				EnsureBinding(StatescriptPropertyDirection.Output, index).Resolver =
-					new VariableResolverResource { VariableName = variableName };
-			}
-		}
-
-		variableDropdown.ItemSelected += x =>
-		{
-			if (NodeResource is null || _graph is null)
-			{
-				return;
-			}
-
-			var oldResolver = FindBinding(StatescriptPropertyDirection.Output, index)?.Resolver?.Duplicate()
-				as StatescriptResolverResource;
-
-			var variableName = _graph.Variables[(int)x].VariableName;
-			var newResolver = new VariableResolverResource { VariableName = variableName };
-			EnsureBinding(StatescriptPropertyDirection.Output, index).Resolver = newResolver;
-
-			if (_undoRedo is not null)
-			{
-				_undoRedo.CreateAction("Change Output Variable", customContext: _graph);
-				_undoRedo.AddDoMethod(
-					this,
-					MethodName.ApplyResolverBinding,
-					(int)StatescriptPropertyDirection.Output,
-					index,
-					(StatescriptResolverResource)newResolver.Duplicate());
-				_undoRedo.AddUndoMethod(
-					this,
-					MethodName.ApplyResolverBinding,
-					(int)StatescriptPropertyDirection.Output,
-					index,
-					oldResolver ?? new StatescriptResolverResource());
-				_undoRedo.CommitAction(false);
-			}
-
-			PropertyBindingChanged?.Invoke();
-		};
-
-		hBox.AddChild(variableDropdown);
-	}
-
-	private void ShowResolverEditorUI(
-		Func<NodeEditorProperty> factory,
-		StatescriptNodeProperty? existingBinding,
-		Type expectedType,
-		VBoxContainer container,
-		StatescriptPropertyDirection direction,
-		int propertyIndex,
-		bool isArray = false)
-	{
-		if (_graph is null)
-		{
-			return;
-		}
-
-		NodeEditorProperty resolverEditor = factory();
-
-		resolverEditor.Setup(
-			_graph,
-			existingBinding,
-			expectedType,
-			() =>
-			{
-				SaveResolverEditorWithUndo(resolverEditor, direction, propertyIndex);
-			},
-			isArray);
-
-		resolverEditor.LayoutSizeChanged += ResetSize;
-
-		container.AddChild(resolverEditor);
-
-		var key = new PropertySlotKey(direction, propertyIndex);
-		_activeResolverEditors[key] = resolverEditor;
-	}
-
-	private void SaveResolverEditorWithUndo(
-		NodeEditorProperty resolverEditor,
-		StatescriptPropertyDirection direction,
-		int propertyIndex)
-	{
-		if (NodeResource is null)
-		{
-			return;
-		}
-
-		StatescriptNodeProperty? existing = FindBinding(direction, propertyIndex);
-		var oldResolver = existing?.Resolver?.Duplicate() as StatescriptResolverResource;
-
-		SaveResolverEditor(resolverEditor, direction, propertyIndex);
-
-		StatescriptNodeProperty? updated = FindBinding(direction, propertyIndex);
-		var newResolver = updated?.Resolver?.Duplicate() as StatescriptResolverResource;
-
-		if (_undoRedo is not null)
-		{
-			_undoRedo.CreateAction("Change Node Property", customContext: _graph);
-			_undoRedo.AddDoMethod(
-				this,
-				MethodName.ApplyResolverBinding,
-				(int)direction,
-				propertyIndex,
-				newResolver ?? new StatescriptResolverResource());
-			_undoRedo.AddUndoMethod(
-				this,
-				MethodName.ApplyResolverBinding,
-				(int)direction,
-				propertyIndex,
-				oldResolver ?? new StatescriptResolverResource());
-			_undoRedo.CommitAction(false);
-		}
-
-		PropertyBindingChanged?.Invoke();
-	}
-
-	private void SaveResolverEditor(
-		NodeEditorProperty resolverEditor,
-		StatescriptPropertyDirection direction,
-		int propertyIndex)
-	{
-		if (NodeResource is null)
-		{
-			return;
-		}
-
-		StatescriptNodeProperty binding = EnsureBinding(direction, propertyIndex);
-		resolverEditor.SaveTo(binding);
-	}
-
 	private void ApplyResolverBinding(
 		int directionInt,
 		int propertyIndex,
@@ -907,216 +585,6 @@ public partial class StatescriptGraphNode : GraphNode
 			{
 				NodeResource.PropertyBindings.RemoveAt(i);
 			}
-		}
-	}
-
-	private void SetupNodeByType(StatescriptNodeType nodeType)
-	{
-		switch (nodeType)
-		{
-			case StatescriptNodeType.Entry:
-				SetupEntryNode();
-				break;
-			case StatescriptNodeType.Exit:
-				SetupExitNode();
-				break;
-			case StatescriptNodeType.Action:
-				SetupActionNode();
-				break;
-			case StatescriptNodeType.Condition:
-				SetupConditionNode();
-				break;
-			case StatescriptNodeType.State:
-				SetupStateNode();
-				break;
-		}
-	}
-
-	private void SetupEntryNode()
-	{
-		CustomMinimumSize = new Vector2(100, 0);
-
-		var label = new Label { Text = "Start" };
-		AddChild(label);
-		SetSlotEnabledRight(0, true);
-		SetSlotColorRight(0, _eventColor);
-
-		ApplyTitleBarColor(_entryColor);
-	}
-
-	private void SetupExitNode()
-	{
-		CustomMinimumSize = new Vector2(100, 0);
-
-		var label = new Label { Text = "End" };
-		AddChild(label);
-		SetSlotEnabledLeft(0, true);
-		SetSlotColorLeft(0, _eventColor);
-
-		ApplyTitleBarColor(_exitColor);
-	}
-
-	private void SetupActionNode()
-	{
-		var label = new Label { Text = "Execute" };
-		AddChild(label);
-		SetSlotEnabledLeft(0, true);
-		SetSlotColorLeft(0, _eventColor);
-		SetSlotEnabledRight(0, true);
-		SetSlotColorRight(0, _eventColor);
-
-		ApplyTitleBarColor(_actionColor);
-	}
-
-	private void SetupConditionNode()
-	{
-		var hBox = new HBoxContainer();
-		hBox.AddThemeConstantOverride("separation", 16);
-		AddChild(hBox);
-
-		var inputLabel = new Label { Text = "Condition" };
-		hBox.AddChild(inputLabel);
-		SetSlotEnabledLeft(0, true);
-		SetSlotColorLeft(0, _eventColor);
-
-		var trueLabel = new Label
-		{
-			Text = "True",
-			HorizontalAlignment = HorizontalAlignment.Right,
-			SizeFlagsHorizontal = SizeFlags.ExpandFill,
-		};
-
-		hBox.AddChild(trueLabel);
-		SetSlotEnabledRight(0, true);
-		SetSlotColorRight(0, _eventColor);
-
-		var falseLabel = new Label
-		{
-			Text = "False",
-			HorizontalAlignment = HorizontalAlignment.Right,
-			SizeFlagsHorizontal = SizeFlags.ExpandFill,
-		};
-
-		AddChild(falseLabel);
-		SetSlotEnabledRight(1, true);
-		SetSlotColorRight(1, _eventColor);
-		ApplyTitleBarColor(_conditionColor);
-	}
-
-	private void SetupStateNode()
-	{
-		var hBox1 = new HBoxContainer();
-		hBox1.AddThemeConstantOverride("separation", 16);
-		AddChild(hBox1);
-
-		var inputLabel = new Label { Text = "Begin" };
-		hBox1.AddChild(inputLabel);
-		SetSlotEnabledLeft(0, true);
-		SetSlotColorLeft(0, _eventColor);
-
-		var activateLabel = new Label
-		{
-			Text = "OnActivate",
-			HorizontalAlignment = HorizontalAlignment.Right,
-			SizeFlagsHorizontal = SizeFlags.ExpandFill,
-		};
-
-		hBox1.AddChild(activateLabel);
-		SetSlotEnabledRight(0, true);
-		SetSlotColorRight(0, _eventColor);
-
-		var hBox2 = new HBoxContainer();
-		hBox2.AddThemeConstantOverride("separation", 16);
-		AddChild(hBox2);
-
-		var abortLabel = new Label { Text = "Abort" };
-		hBox2.AddChild(abortLabel);
-		SetSlotEnabledLeft(1, true);
-		SetSlotColorLeft(1, _eventColor);
-
-		var deactivateLabel = new Label
-		{
-			Text = "OnDeactivate",
-			HorizontalAlignment = HorizontalAlignment.Right,
-			SizeFlagsHorizontal = SizeFlags.ExpandFill,
-		};
-
-		hBox2.AddChild(deactivateLabel);
-		SetSlotEnabledRight(1, true);
-		SetSlotColorRight(1, _eventColor);
-
-		var abortOutputLabel = new Label
-		{
-			Text = "OnAbort",
-			HorizontalAlignment = HorizontalAlignment.Right,
-			SizeFlagsHorizontal = SizeFlags.ExpandFill,
-		};
-
-		AddChild(abortOutputLabel);
-		SetSlotEnabledRight(2, true);
-		SetSlotColorRight(2, _eventColor);
-
-		var subgraphLabel = new Label
-		{
-			Text = "Subgraph",
-			HorizontalAlignment = HorizontalAlignment.Right,
-			SizeFlagsHorizontal = SizeFlags.ExpandFill,
-		};
-
-		AddChild(subgraphLabel);
-		SetSlotEnabledRight(3, true);
-		SetSlotColorRight(3, _subgraphColor);
-
-		ApplyTitleBarColor(_stateColor);
-	}
-
-	private void ClearSlots()
-	{
-		foreach (Node child in GetChildren())
-		{
-			RemoveChild(child);
-			child.QueueFree();
-		}
-	}
-
-	private void ApplyTitleBarColor(Color color)
-	{
-		var titleBarStyleBox = new StyleBoxFlat
-		{
-			BgColor = color,
-			ContentMarginLeft = 12,
-			ContentMarginRight = 12,
-			ContentMarginTop = 6,
-			ContentMarginBottom = 6,
-			CornerRadiusTopLeft = 4,
-			CornerRadiusTopRight = 4,
-		};
-
-		AddThemeStyleboxOverride("titlebar", titleBarStyleBox);
-
-		var selectedTitleBarStyleBox = (StyleBoxFlat)titleBarStyleBox.Duplicate();
-		selectedTitleBarStyleBox.BgColor = color.Lightened(0.2f);
-		AddThemeStyleboxOverride("titlebar_selected", selectedTitleBarStyleBox);
-	}
-
-	private void ApplyBottomPadding()
-	{
-		StyleBox? existing = GetThemeStylebox("panel");
-
-		if (existing is not null)
-		{
-			var panelStyle = (StyleBox)existing.Duplicate();
-			panelStyle.ContentMarginBottom = 10;
-			AddThemeStyleboxOverride("panel", panelStyle);
-		}
-
-		StyleBox? selectedExisting = GetThemeStylebox("panel_selected");
-
-		if (selectedExisting is not null)
-		{
-			var selectedPanelStyle = (StyleBox)selectedExisting.Duplicate();
-			selectedPanelStyle.ContentMarginBottom = 10;
-			AddThemeStyleboxOverride("panel_selected", selectedPanelStyle);
 		}
 	}
 }
